+++ src/main/java/com/agnieszka/reservationapp/vetapp/repository/DoctorRepository.java
@@ -3,7 +3,6 @@ package com.agnieszka.reservationapp.vetapp.repository;
 import com.agnieszka.reservationapp.vetapp.model.Doctor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,8 +12,8 @@ public interface DoctorRepository  {
 
     Page<Doctor> findAll(Pageable page);
 
-    Optional<Doctor> findById(Integer id);
+    Optional<Doctor> findById(Long id);
 
-    boolean existsById(Integer id);
+    boolean existsById(Long id);
     Doctor save(Doctor entity);
 }
